// src/types/vision.ts
export type Vision = {
  id: string;

  section_heading: string;
  main_heading: string;
  description: string;

  usp_1_title: string;
  usp_1_description: string;

  usp_2_title: string;
  usp_2_description: string;

  usp_3_title: string;
  usp_3_description: string;

  usp_4_title: string;
  usp_4_description: string;

  closing_statement: string;

  image_url: string | null;

  created_at: string;
  updated_at: string;
};